完善了jmc_dtos_ex()函数NaN和+/-Inf的输出.

+++ src/jimic/string/sprintf.inl.h
@@ -411,14 +411,14 @@ vslprintf_out_double:
                     //    && align == FMT_ALIGN_DEFAULT && fill == FMT_FILL_DEFAULT) {
                     //if ((flag | fill) == (FMT_DEFAULT_FLAG | FMT_ALIGN_DEFAULT | FMT_FILL_DEFAULT)) {
                     if ((flag | fill) == FMT_DEFAULT_STATUS) {
-                        len = jmc_dtos(buf, dbl, width, precision);
-                        //len = jmc_dtos_ex(buf, -1, dbl, flag, fill, width, precision);
+                        len = jmc_dtos(buf, dbl, (int)width, precision);
+                        //len = jmc_dtos_ex(buf, -1, dbl, flag, fill, (int)width, precision);
                         buf += len;
                         cur++;
                         goto vslprintf_try_next;
                     }
                     else {
-                        len = jmc_dtos_ex(buf, -1, dbl, flag, fill, width, precision);
+                        len = jmc_dtos_ex(buf, -1, dbl, flag, fill, (int)width, precision);
                         buf += len;
                         cur++;
                         goto vslprintf_try_next;
